+++ modules/visual-tests/tests/tests.spec.ts
@@ -112,6 +112,7 @@ import { testing as TagPill__test } from '../src/routes/test/TagPill';
 import { testing as TextBlock__test } from '../src/routes/test/TextBlock';
 import { testing as TextButton__test } from '../src/routes/test/TextButton';
 import { testing as TextInput__test } from '../src/routes/test/TextInput';
+import { testing as Tooltip__test } from '../src/routes/test/Tooltip';
 import { testing as VSpacer__test } from '../src/routes/test/VSpacer';
 import { testing as WizardLayout__test } from '../src/routes/test/WizardLayout';
 import { testing as WizardStepper__test } from '../src/routes/test/WizardStepper';
@@ -201,6 +202,7 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   TextBlock: TextBlock__test,
   TextButton: TextButton__test,
   TextInput: TextInput__test,
+  Tooltip: Tooltip__test,
   VSpacer: VSpacer__test,
   WizardLayout: WizardLayout__test,
   WizardStepper: WizardStepper__test,